% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_pdf_report}
\alias{create_pdf_report}
\title{create_pdf_report}
\usage{
create_pdf_report(result, pdf_filename)
}
\arguments{
\item{result}{ReactomeAnalysisResult. The result to create the report for.}

\item{pdf_filename}{character. Path where the PDF file will be written to. This file
will be overwritten if it exists}
}
\description{
Creates a PDF report of a ReactomeGSA result. This function
currently only supports GSA-based analysis results.
}
\examples{
library(ReactomeGSA.report)
library(ReactomeGSA.data)
data(griss_melanoma_result)

pdf_file_path <- tempfile(pattern = "reactome_report_", tmpdir = tempdir(), fileext = ".pdf")

# create the PDF report at the specified location (simply a temporary file in this example)
create_pdf_report(result = griss_melanoma_result, pdf_filename = pdf_file_path)
}
